Retry certain operations until they succeed (fixes #29)
