+++ ci/travis/script.sh
@@ -37,7 +37,7 @@ echo "TRAVIS_BRANCH=$TRAVIS_BRANCH"
 echo "TRAVIS_PULL_REQUEST=$TRAVIS_PULL_REQUEST"
 
 julia $JULIA_FLAGS -e 'import Pkg; Pkg.resolve();'
-julia $JULIA_FLAGS deps/build.jl
+julia $JULIA_FLAGS -e 'import Pkg; Pkg.build("MirrorUpdater");'
 julia $JULIA_FLAGS run-mirror-updater.jl --delete-gists-older-than-minutes 10080 --gist-description "$GIST_DESCRIPTION" --task "$TASK" $DRY_RUN
 
 cat Project.toml
