+++ src/MirrorUpdater.jl
@@ -4,6 +4,17 @@ module MirrorUpdater # Begin module MirrorUpdater
 
 __precompile__(true)
 
+import ArgParse
+import Conda
+import Coverage
+import Dates
+import GitHub
+import HTTP
+import JSON
+import Pkg
+import Test
+import TimeZones
+
 include(joinpath("package_directory.jl"))
 include(joinpath("version.jl"))
 include(joinpath("welcome.jl"))
